pyproject.toml
[tool.poetry]
name = "llm-tools"
version = "0.1.0"
description = ""
authors = ["n-kats <n-kats19890214@hotmail.co.jp>"]
readme = "README.md"
packages = [{include = "llm_clis"}, {include = "llm_tools"}, {include = "voice_browser"}]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
openai = ">=1.6.1"
click = "^8.1.3"
websockets = "^11.0.3"
sqlalchemy = "^2.0.16"
sqlalchemy-serializer = "^1.4.1"
fastapi = "^0.100.0"
pydantic = "^2.0.3"
uvicorn = "^0.23.1"
arxiv = "^1.4.8"
yq = "^3.2.2"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
pydrive2 = "^1.17.0"
google-api-python-client = "^2.103.0"
google-auth-httplib2 = "^0.1.1"
google-auth-oauthlib = "^1.1.0"
oauth2client = "^4.1.3"
pydub = "^0.25.1"
tenacity = "^8.2.3"
jinja2 = "^3.1.2"
beautifulsoup4 = "^4.12.2"


[tool.poetry.group.dev.dependencies]
ipython = "^8.14.0"
black = "^23.3.0"
isort = "^5.12.0"
pdbpp = "^0.10.3"
ruff = "^0.1.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
chat-store-watch = "llm_clis.chat_store.cli:watch"
chat-store = "llm_clis.chat_store.cli:send"
text-to-voice = "llm_clis.text_to_voice.main:main"

[[tool.mypy.overrides]]
module=[
  "yaml", "arxiv", "dotenv",
  "bs4", "requests",
  "click", "fastapi.*", "websockets.*",
  "sqlalchemy.*","sqlalchemy_serializer",
  "pydub",
  "google.*", "googleapiclient.*"
]
ignore_missing_imports = true
